+++ .changeset/fast-knives-peel.md
@@ -1,12 +1,9 @@
 ---
-"@clerk/astro": patch
 "@clerk/backend": patch
-"@clerk/express": patch
-"@clerk/fastify": patch
-"@clerk/nextjs": patch
-"@clerk/remix": patch
-"@clerk/clerk-sdk-node": patch
-"@clerk/types": patch
 ---
 
-Move `AuthObject` and its related types to `@clerk/types`. You will now be able to import it from there.
+Export the type `AuthObject`. You can now use it like so:
+
+```ts
+import type { AuthObject } from "@clerk/backend"
+```
