+++ .changeset/itchy-ravens-protect.md
@@ -1,0 +1,10 @@
+---
+"@clerk/clerk-sdk-node": patch
+"@clerk/express": patch
+"@clerk/fastify": patch
+"@clerk/nextjs": patch
+"@clerk/astro": patch
+"@clerk/remix": patch
+---
+
+Internal change: Use `AuthObject` type import from `@clerk/backend`.
