+++ packages/fastify/src/getAuth.ts
@@ -1,4 +1,4 @@
-import type { AuthObject } from '@clerk/types';
+import type { AuthObject } from '@clerk/backend';
 import type { FastifyRequest } from 'fastify';
 
 import { pluginRegistrationRequired } from './errors';
