+++ packages/remix/src/ssr/types.ts
@@ -1,7 +1,6 @@
-import type { Organization, Session, User, VerifyTokenOptions } from '@clerk/backend';
+import type { AuthObject, Organization, Session, User, VerifyTokenOptions } from '@clerk/backend';
 import type { RequestState } from '@clerk/backend/internal';
 import type {
-  AuthObject,
   LegacyRedirectProps,
   MultiDomainAndOrProxy,
   SignInFallbackRedirectUrl,
